+++ docs/13_0_0/gettingstarted/whatsnew.md
@@ -74,6 +74,7 @@ This page contains a list of big features. Please check the GitHub issues for al
     * Added attribute `ariaLabel` to allow screen reader support on button and `title` for tooltip.
 * SplitButton
     * Added attribute `filterNormalize` to allow normalized filtering (without accents).
+    * Added attribute `filterInputAutoFocus` to control filter input should receive focus when overlay popup is displayed.
     * `OverlayPanel` can be added as child component to use as the overlay instead of `MenuItem`s.
 * StaticMessage
     * Added attribute `display` with supported values `both` (default), `icon` and `text`.